# Rework request's logs