Adds log line with received new request's information, that includes:
- HTTP method
- URL
- local and remote endpoints
- headers in form of {"Header1: Value1", "Header2: Value2", ...}, values are escaped

The log line is printed on INFO log level.