@shindo 2015-08-06
Related issue #57 requires also the following:
- markers between keep-alive requests
- additional information in case of bad or malformed request

I'll check if it's really needed and update the pull request.