@shindo 2015-08-06
@iderikon, Thanks, I've updated the pull request with your suggestions.