@shindo 2015-08-06
'Process next request' log line is printed on INFO log level now. Local and remote endpoints added to the log line.