@shindo 2015-08-06
To obtain additional information in case of bad or malformed request it's required to update the request parser to produce error message about invalid request.
Right now it just responds with `false` and TheVoid server responds with bad request with no additional information in the log.

This will be done in a separate pull request.