@shindo 2015-08-10
@bioothod, I've updated the pull request according to your comments. Thanks for the valuable feedback.

I looks like Boost.Spirit is not thread-safe by default, but with the following define it's possible to use grammar in multithreaded code ([Spirit Docs](http://www.boost.org/doc/libs/1_46_1/libs/spirit/classic/doc/grammar.html), section "Reentrancy and multithreading"):
```(cpp)
#define BOOST_SPIRIT_THREADSAFE
```

Still, Boost.Spirit uses mutexes to protect access to shared variables, which is unfortunate.

`log_request_headers` config option is added. This option accepts a vector of strings (names of request's headers to log), and it's empty by default.

String escaping is based on python's string representation (function `PyString_Repr`):
http://svn.python.org/projects/python/trunk/Objects/stringobject.c