@bioothod 2015-08-11
This looks really good, thank you. I only have couple of comments: docs and perf test to understand the influence, and it is actually ready to be merged