@shindo 2015-08-11
I've updated example server's config and documentation:
- [commit](https://github.com/shindo/swarm/commit/24ad671b5aba6162822d3823216a5994b8195207)
- http://thevoid.readthedocs.org/en/docs/server.html#configuration