@shindo 2015-08-14
I've made performance testing with [ab](http://httpd.apache.org/docs/2.2/programs/ab.html) (one may install it on Ubuntu with apache2-utils package).

[Test server](https://github.com/reverbrain/swarm/blob/master/tests/server.cpp) with [ok handler](https://github.com/reverbrain/swarm/blob/master/tests/handlers/ok.cpp) was used with the following configuration:
```(javascript)
{
    "endpoints": [
        "0.0.0.0:8080"
    ],
    "backlog": 512,
    "threads": 2,
    "buffer_size": 65536,
        "logger": {
        "level": "info",
        "frontends": [
            {
                "formatter": {
                    "type": "string",
                    "pattern": "%(timestamp)s %(request_id)s/%(lwp)s/%(pid)s %(severity)s: %(message)s, %(...L)s"
                },
                "sink": {
                    "type": "files",
                    "path": "/home/shindo/repos/swarm/build/tests/server.log",
                    "autoflush": true,
                    "rotation": { "move": 0 }
                }
            }
        ]
    },
    "daemon": {
        "fork": false,
        "uid": 1000
    },
    "monitor-port": 20000,
    "log_request_headers": ["Authorization", "Cookie", "User-Agent"],
    "application": {
        "handlers": [
            {
                "handler": "ok",
                "prefix_match": "/ping"
            }
        ]
    }
}
```

The following command was used to run **ab**:
```(bash)
ab -n 150000 -m GET -H 'Content-Type: application/json' -H 'Authorization: Basic ZnJlZDpmcmVk' -H 'Cookie: AAA...AAA' http://0.0.0.0:8080/ping
```

Here:
- 150000 requests with concurrency = 1 (number of multiple requests to make at a time)
- 'Cookie' header's size is 5Kb

Results without headers dump (TheVoid's master branch):
```
Server Hostname:        0.0.0.0
Server Port:            8080

Document Path:          /ping
Document Length:        0 bytes

Concurrency Level:      1
Time taken for tests:   90.255 seconds
Complete requests:      150000
Failed requests:        0
Total transferred:      8550000 bytes
HTML transferred:       0 bytes
Requests per second:    1661.96 [#/sec] (mean)
Time per request:       0.602 [ms] (mean)
Time per request:       0.602 [ms] (mean, across all concurrent requests)
Transfer rate:          92.51 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.0      0       1
Processing:     0    1   0.8      0      79
Waiting:        0    0   0.7      0      77
Total:          0    1   0.8      0      79
WARNING: The median and mean for the processing time are not within a normal deviation
        These results are probably not that reliable.
WARNING: The median and mean for the total time are not within a normal deviation
        These results are probably not that reliable.

Percentage of the requests served within a certain time (ms)
  50%      0
  66%      1
  75%      1
  80%      1
  90%      1
  95%      1
  98%      2
  99%      3
 100%     79 (longest request)
```
Total server's log size: 86M

Results with headers dump:
```
Server Hostname:        0.0.0.0
Server Port:            8080

Document Path:          /ping
Document Length:        0 bytes

Concurrency Level:      1
Time taken for tests:   136.187 seconds
Complete requests:      150000
Failed requests:        0
Total transferred:      8550000 bytes
HTML transferred:       0 bytes
Requests per second:    1101.42 [#/sec] (mean)
Time per request:       0.908 [ms] (mean)
Time per request:       0.908 [ms] (mean, across all concurrent requests)
Transfer rate:          61.31 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.0      0       2
Processing:     1    1   1.2      1     181
Waiting:        0    1   1.1      1     181
Total:          1    1   1.2      1     181

Percentage of the requests served within a certain time (ms)
  50%      1
  66%      1
  75%      1
  80%      1
  90%      1
  95%      1
  98%      3
  99%      3
 100%    181 (longest request)
```
Total server's log size: 842M

Thus we can see that performance degrade by 30%, mainly because of the resulting log size as with empty `log_request_headers` we get:
```
Requests per second:    1536.94 [#/sec] (mean)
Time per request:       0.651 [ms] (mean)
```
Total server's log size: 115M

Still I'm not sure if we can talk here about percentages as 0.602ms to 0.651ms degradation is ~10%, but it's just 50us in absolute numbers.